trivial renamings
