+++ src/MyLambdaDotNetCoreProject.Api/MyLambdaDotNetCoreProject.Api.csproj
@@ -38,4 +38,7 @@
       <CopyToOutputDirectory>Never</CopyToOutputDirectory>
     </None>
   </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Base\" />
+  </ItemGroup>
 </Project>
