+++ src/MyLambdaDotNetCoreProject.Application/Queries/IQuery.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace MyLambdaDotNetCoreProject.Application.Queries
+{
+    public interface IQuery
+    {}
+}
